Fix broken combinators library
An attempt at defining the combinator `and_then` for Option is currently being
made.  This is incorrect, the code still compiles because the function is marked
`pub fn` (within a library).

The `and_then` combinator is implemented for the Option<T> type by the standard
library, the call we make is calling this implementation _not_ the function we
define.  We should just remove the definition from our library.

Also

 - Rename function `extension` -> `suffix`
 - Improve funciton argument name
 - Use `match` explicitily instead of `?` operator to demonstrate it usage